refactor: split date utils up into separate files
